partly implemented search fragment logic

+++ remote/src/main/java/com/example/remote/models/search/ThreeHours.kt
@@ -5,7 +5,7 @@ data class ThreeHours(
     val dt: Int?,
     val dt_txt: String?,
     val main: Main?,
-    val pop: Int?,
+    val pop: Double?,
     val rain: Rain?,
     val snow: Snow?,
     val sys: Sys?,
